WebStorage/tests/approved/storage_local_getitem_null.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_local_getItem_null</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('localStorage' in window) && window.localStorage !== null){
			localStorage.clear();
			localStorage.setItem("name", "user1");
			localStorage.setItem("age", "20");

			try {
				test(function() {
					assert_equals(localStorage.getItem("unknown"), null, "localStorage.getItem('name')")
				}, "getItem method test")
			} catch (e) {
				test(function()
				{
					assert_unreached("getItem method is failed.");
				}, "getItem method test")
			}
		} else {
			test(function()
			{
				assert_unreached("localStorage is not supported.");
			}, "localStorage support test")
		}	
	</script>

 </body>
</html>
